EREF-64 Machine readable test specification solution for identifiers query (#56)
* EREF-64 add schema and example data for identifier-query-test-cases

* EREF-64 add Java test that implements identifier-query-test-cases

Note that the schema identifier-query-test-cases.xsd refers to such
types that are already used for code generation in the commons module.
We want to use those commons module classes but there is no way to
tell code generation to not generate said classes. Let's just remove
the duplicate classes after code generation.

* EREF-64 identifier-query-test-cases omits irrelevant values

Values that have no relevance with regard to query logic are omitted
in test cases. If implementation requires any of the omitted values
then it should populate those with random values.

* EREF-64 IdentifiersQueryTest populates required values omitted in specs