EREF-27 eFTI Platform identifiers to eFTI Gate REST API interface (#72)
* EREF-27 allow access to gate's platform REST api via pre-auth headers

Gate expects to get headers that identify the calling client and its
role. These headers must be added by some other component, that is
responsible for authenticating the client. This component may be,
for example, a proxy like nginx.

For local development, configure the Apache proxy to accept headers
X-Mock-Pre-Authenticated-User-Id and X-Mock-Pre-Authenticated-User-Role
that may be used to authenticate the client.

* EREF-27 gate's platform REST api supports uploading identifiers

* EREF-27 platform simulator calls gate whoami at startup

* EREF-27 platform simulator supports consignment upload via REST api

Add new platform simulator api /upload/consignment/{datasetId}
where a common consignment xml may be uploaded. The xml is saved
on disk and converted to identifiers xml which is uploaded to gate.

* EREF-27 add consignment upload to postman collection

* EREF-27 setup integration test with testcontainers, test api security

* EREF-27 setup database support for integration tests

Database is cleared before each test.

* EREF-27 add integration tests for platform api

* EREF-27 use reproducible pseudo-random data in integration tests

The idea is that dependencies on specific data values do not creep into
tests. The generated data is reproducible: before each test, the
random seed is reset. The seed is based on current date and is logged.

* EREF-27 fix platform-gate-simulator compilation in IDE

Generated test sources were marked as prod sources in IDEA.

* EREF-27 add error handling tests, minor refactoring

* EREF-27 platform-simulator saves xml, add module test-support

Move test data utilities to test-support module so that they may be
used by different modules.